Unify fields in backupBase struct
Use only a single struct to track both the snapshot and backup for a
given corso backup. This makes some logistics easier when dealing with
struct internals.

Leave existing functions in place to reduce test churn. Eventually
though it would be good to return just the struct containing both items.