Basic skeleton of directory iterator
Create the required functions for a directory iterator implementation.
As corso has multiple data sources for a single directory returned to
kopia use a state machine architecture to handle
setup/teardown/switching between the different data sources.

Inner functions to actually get a single entry to return to kopia will
be implemented in the next commit.