Fail backups if we see an item has a zero mod time
This is a big hammer solution to keep bad state from getting into the
repo. It will keep us getting into a situation in which we've made a
backup that was marked as a successful backup (and thus a merge base)
but has items which we won't be able to do details merging on if we do
use it as a merge base.

Long term, we should probably tag these backups with some sort of
special tag that says they shouldn't be used as bases at all since we'll
fail to merge details. We'll fail details merging even if the base is
used as an assist base.