start centralizing intg test setup
we duplicate a lot of boilerplate code in integration tests
involving the creation of an api client and the aggregation
of external IDs.  In some packages we've begun
to add quality-of-life structs to aggregate these setup details.
This change begins moving those structs to a singular
package so that we can clear out the common setup boilerplate
found elsewhere in the code.