Rename all existing export handlers
Rename these handlers to be base<Service>Handler so that we get some
segmentation of functionality. This will allow us a bit of compile-time
safety when it comes to accessing graph because the base handler doesn't
contain a client instance. Essentially it allows us local access to
things for operations like restore.

Future additions to the handler that require a client should wrap this
base handler to provide that functionality. Export functions shouldn't
be updated to be part of the new handler wrapper but should stay as part
of the base handler so they continue to not have access to graph.