Rework getting results; add cancellation
Rework how a pager finishes things so it can be cancelled and will exit
sooner and more reliably. The previous method of draining the results
channel could lead to cases where the pager would continue fetching data
even after the context was cancelled because the select-block picks a
random channel that is able to send/receive.

The new implementation adds a couple more channels so we can both cancel
the pager and check when the pager is done without having to worry about
getting more pages.