Add options to limit data backed up
These are a bit hacky, but since it's not exposed directly to CLI users
and we have a bit more leeway with SDK users, it gives us some wiring to
set limits for preview backups. These values won't be persisted so we
can streamline them as needed.

The unfortunately thing about trying to set options for preview backups
is they can become arbitrarily complex depending on how precise you want
them to be. This takes the "quick and easy" approach by just applying
the limits to all data categories in the backup independently. See the
comment for more details.