clean up variable naming and usage in collections test (#4648)
the collections_test.go contains a lot of hardcoded string references,
and a jumble of ways to use them. Furthermore, it was impelemented from
the beginning with many overlapping strings that are used both as IDs
and non-id properties, which causes significant confusion for new
maintainers.

The goals of this PR are the following:
1. remove common hardcoded string values
2. make id and non-id value cases clearly visible
3. centralize path production so path structure is easier to grok
4. make references uniform throughout the file.

the result is a code set that's a little more difficult to read at first
blush, but significantly easier to understand and maintain without
needing to dive deep into the test structure.

No logical changes, only renames/code movement.