move service api to client methods
sdk consumers need a way to configure the graph api client
options when using the services api.  This is the first in a two-part
change to expose those options.  This step moves the api to a
client-based set of methods instead of free functions.  The next
step will add client configuration to the service client.