add new funcs and tidy up drive limiter
adds some new functions to the drive limiter that will
be used specifically in the tree-based backup process.
Also updates the limiter tests to have separate versions
for the tree and non-tree variations.  In this PR, the tree
variation will definitely fail.  The next PR will be focused on
backup process corrections needed to ensure that the limit
handling is correct according to the existing tests.