swap stats for in-tree item counts for limits
In the tree-based implementation of drive collections.Get,
swaps out the driveEnumerationStats and instead uses
the current state of the tree to count the items within
and compare those counts to the expected preview limits.

Tests are known to still be failing.  Next PR will bring things back to green.