add multiple delta iteration to getTree (#4747)
expands the drive item enumeration by running multiple delta queries per
backup, with the intent to reach a no-change state in the given delta.
Most of the code here modifies the test structures to accomodate the
possibility of a multi-delta enumeration within the tree version. Plus
renaming and other refactoring for quality of life. Full test coverage
of the change won't be included until the next PR.