generate collectables from tree (#4779)
post-process handling to produce a set of collectable data from the
current state of the tree. Collectable data is a minimization of
collection data, where the collection will require additional properties
that aren't owned by the tree. The caller is expected to be able to
produce a complete data.BackupCollection and newPrevPaths from the
produced collectables.