Fail test if fields are unpopulated
To ensure the test continues to check what we want as the code changes
going forward use the reflect package to walk all fields in the config
struct and make sure none of them have the zero value. If a zero-valued
field is found, print the set of field indices used to reach it and fail
the test.