Update test to check sanitized input
Test deserializing and sending the sanitized version of the output from
the original test. Also ensure that the returned body from the sanitized
input is the same as the body from the original input.

We compare only the body instead of the entire serialized content
because the item's ID, mod time, etc will change between the versions.