centralize core err interpretation
now that we have a core error library, we can start swapping out
the percolation of graph api errors with core error sentinels.  This PR
begins that process by adding a centralized transformer to the
graph service handlers that ensures all calls produce the same set
of core errors according to some standard categorizations.
This is just an initial step, and does not transfer all graph errors
into core sentinels.