deduplicate graph api errors (#4973)
the graph api service wrapper is currently stacking every retried error,
leading to errors that read like, "foo; bar; baz; baz; baz". This output
is confusing and unhelpful for debugging. The changes in this PR reduce
that duplication while still adding the prior error cases into the
returned context so that the returned errors are easier to grok.