remove resourcenotfound error in place of notfound
resourceNotFound is an interpretation of the generic NotFound error.  It's presence causes some
unwanted confusion, like, what's the difference between the two errors?  Or: are both errors equally
viable for endpoints that fetch resources?

Removing resourceNotFound simplifies error categorization.  It's up to the caller to
qualify the thing that's not found as a resource or some other entity.