Pick behavior for edge case during restore
A few things make this necessary:
  * it's possible for a user to have a folder called "contacts" or
    "Contacts"
  * trying to make a folder called "contacts" or "Contacts" creates a
    subfolder at the root instead of returning already exists
  * Corso flattens the contacts hierarchy so LocationRefs only contain
    the leaf folder

Given the above, the following is true:
  * during restore we can't tell if we have the root folder or a
    Contacts/Contacts folder since we only get the LocationRef
  * any in-place restore to the root folder requires a special case for
    the contacts root else we'll end up with subfolders

This patch goes the route of merging contacts in the root folder and any
folders named the same as the root. Since we use an English locale
string for the root that's the only one that will be merged if English
locale was used during the original backup.