Update repo init for partial init case
Update the repo init logic to continue if the repo is marked as already
existing. If it already exists then the process will try to update
persistent config info in the repo. If the repo already exists but some
connection credential is incorrect then stack errors such that both the
reason for the connect failure and the fact that the repo already exists
are both visible.