Add a pin to kopia snapshots so they won't be GCed
Internally kopia has the notion of a retention policy. This adds a pin
to each snapshot that corso uses kopia to create so that those snapshots
won't be subject to kopia's rentention policy. As we disable kopia's
retention policy during repo init this just acts as another layer of
security to ensure snapshots won't be deleted unexpectedly.