add authentication to requester
the graph requester for large item downloads now includes
the option to authenticate requests.  The option is configured
at the time of creating the requester, therefore all requests
using that servier are either authenticatd or not.
In our case, we're opting to authenticate all requests, since we
do not use this requester for non-graph api calls, and even
if we did the addition of auth headers is likely benign.