Check for and retry 404s with no content
We've started seeing 404 errors with no content being returned. Check
for these in the http wrapper we use and retry them.

While graph SDK returns an error for this sort of situation it's a very
basic error since it normally expects to parse info out of the response
body. Therefore it should be safe to inject our own error that we can
check for.